requirements.txt
aiocoap==0.4.7
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
apturl==0.5.2
astral==2.2
async-timeout==4.0.2
atomicwrites-homeassistant==1.4.1
attrs==22.2.0
awesomeversion==22.9.0
bcrypt==4.0.1
blinker==1.4
Brlapi==0.8.3
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
ciso8601==2.3.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==40.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
eltakobus @ git+https://github.com/grimmpp/eltako14bus@master
enocean==0.60.1
frozenlist==1.3.3
h11==0.14.0
home-assistant-bluetooth==1.10.0
homeassistant==2023.5.3
httpcore==0.17.0
httplib2==0.20.2
httpx==0.24.0
idna==3.3
ifaddr==0.1.7
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lru-dict==1.1.8
macaroonbakery==1.3.1
MarkupSafe==2.1.2
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
orjson==3.8.10
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.1.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
pyserial-asyncio==0.6
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-slugify==4.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
reportlab==3.6.8
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
ssh-import-id==5.11
systemd-python==234
termcolor==2.3.0
text-unidecode==1.3
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ulid-transform==0.7.2
unattended-upgrades==0.1
urllib3==1.26.5
voluptuous==0.13.1
voluptuous-serialize==2.6.0
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
